Move MastForestView into its own file
